trends/requirements.txt
pytrends==4.9.2
kafka-python-ng
pandas
numpy